requirements.txt
pi==0.1.2
pillow==12.1.1
wonderwords==3.0.1
